# API-Gateway/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.ErrorResponse:
    description: Расширенный ответ с ошибкой
    properties:
      code:
        example: INVALID_INPUT
        type: string
      details:
        items:
          $ref: '#/definitions/models.ValidationError'
        type: array
      error:
        example: Bad Request
        type: string
      message:
        example: Invalid input data
        type: string
    type: object
  models.FileInfo:
    properties:
      category:
        description: '"avatar", "resume", "logo", etc.'
        type: string
      direct_url:
        description: Прямая ссылка через бекенд
        type: string
      id:
        type: string
      name:
        type: string
      type:
        description: '"image", "document", "other"'
        type: string
      url:
        description: Presigned URL для фронта
        type: string
    type: object
  models.FileUploadResponse:
    properties:
      file_info:
        $ref: '#/definitions/models.FileInfo'
      message:
        type: string
    type: object
...
  models.ValidationError:
    description: Детализированная ошибка валидации
    properties:
      code:
        example: REQUIRED_FIELD
        type: string
      field:
        example: email
        type: string
      message:
        example: Email is required
        type: string
    type: object
host: localhost:8000
info:
  contact:
    email: support@studjobs.ru
    name: API Support
  description: API Gateway для платформы StudJobs HH
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: StudJobs HH API Gateway
  version: "1.0"
paths:
...
  /company/{id}/files/documents:
    post:
      consumes:
      - multipart/form-data
      description: Загружает документ для компании текущего пользователя
      parameters:
      - description: ID компании
        in: path
        name: id
        required: true
        type: string
      - description: Файл документа (макс. 20MB)
        in: formData
        name: document
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Информация о загруженном файле
          schema:
            $ref: '#/definitions/models.FileUploadResponse'
        "400":
          description: Неверный запрос или файл слишком большой
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Неавторизованный доступ
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Доступ запрещен
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Загрузить документ компании
      tags:
      - Files
  /company/{id}/files/logo:
    delete:
      consumes:
      - application/json
      description: Удаляет логотип компании текущего пользователя
      parameters:
      - description: ID компании
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Сообщение об успешном удалении
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: У компании нет логотипа
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Неавторизованный доступ
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Доступ запрещен
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Компания не найдена
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удалить логотип компании
      tags:
      - Files
    post:
      consumes:
      - multipart/form-data
      description: Загружает логотип для компании текущего пользователя
      parameters:
      - description: ID компании
        in: path
        name: id
        required: true
        type: string
      - description: Файл логотипа (макс. 5MB)
        in: formData
        name: logo
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Информация о загруженном файле
          schema:
            $ref: '#/definitions/models.FileUploadResponse'
        "400":
          description: Неверный запрос или файл слишком большой
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Неавторизованный доступ
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Доступ запрещен
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Загрузить логотип компании
      tags:
      - Files
...
  /files/{entity_id}/{file_name}:
    get:
      consumes:
      - application/json
      description: Перенаправляет на presigned URL для скачивания файла
      parameters:
      - description: ID сущности (пользователь, компания, вакансия)
        in: path
        name: entity_id
        required: true
        type: string
      - description: Имя файла
        in: path
        name: file_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "302":
          description: Перенаправление на URL файла
          schema:
            type: string
        "404":
          description: Файл не найден
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Прямая загрузка файла
      tags:
      - Files
...
  /users/files/avatar:
    delete:
      consumes:
      - application/json
      description: Удаляет аватар текущего пользователя
      produces:
      - application/json
      responses:
        "200":
          description: Сообщение об успешном удалении
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: У пользователя нет аватара
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Неавторизованный доступ
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удалить аватар пользователя
      tags:
      - Files
    post:
      consumes:
      - multipart/form-data
      description: Загружает аватар для текущего пользователя
      parameters:
      - description: Файл аватара (макс. 5MB)
        in: formData
        name: avatar
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Информация о загруженном файле
          schema:
            $ref: '#/definitions/models.FileUploadResponse'
        "400":
          description: Неверный запрос или файл слишком большой
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Неавторизованный доступ
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Загрузить аватар пользователя
      tags:
      - Files
  /users/files/resume:
    delete:
      consumes:
      - application/json
      description: Удаляет резюме текущего пользователя
      produces:
      - application/json
      responses:
        "200":
          description: Сообщение об успешном удалении
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: У пользователя нет резюме
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Неавторизованный доступ
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удалить резюме пользователя
      tags:
      - Files
    post:
      consumes:
      - multipart/form-data
      description: Загружает резюме для текущего пользователя
      parameters:
      - description: Файл резюме (макс. 10MB)
        in: formData
        name: resume
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Информация о загруженном файле
          schema:
            $ref: '#/definitions/models.FileUploadResponse'
        "400":
          description: Неверный запрос или файл слишком большой
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Неавторизованный доступ
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Загрузить резюме пользователя
      tags:
      - Files
...
securityDefinitions:
  BearerAuth:
    description: 'JWT токен в формате: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
  RoleAuth:
    description: Роль пользователя
    in: header
    name: X-User-Role
    type: apiKey
swagger: "2.0"